Migrate visualization to top level namespace (#1878)
* Migrate visualization to top level namespace

This commit moves the visualization code out of tools into it's own top
level namespace, 'qiskit.visualization'. The visualization was a bit out
of place in the tools directory and is really it's own top level thing
at this point. However, because we declared the public visualization
interfaces stable as part of the 0.7 release we must maintain backwards
compatibility. In order to facilitate that all the public APIs are still
accessible via 'qiskit.tools.visualization'. However, private
api/modules are not accessible off of 'qiskit.tools.visualization'
anymore.

* Add changelog

* Fix lint

Sigh pylint, raising an error for both using a wildcard import and not
using all the imports from the wild card.

* Rebase updates

* Move tests to match new dir structure

* Update test_visualization import per review comment

* Remove underscores from visualization modules

* Fix lint
