+++ qiskit/tools/visualization.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+# pylint: disable=unused-wildcard-import,missing-docstring,wildcard-import
+
+# NOTE(mtreinish): Import public api here to maintain backwards compat
+from qiskit.visualization import *
