+++ qiskit/visualization/bloch.py
