+++ qiskit/visualization/exceptions.py
