+++ qiskit/visualization/interactive/_iplot_blochsphere.py
@@ -14,7 +14,7 @@ import time
 import re
 import numpy as np
 from qiskit.quantum_info import Pauli
-from qiskit.tools.visualization._utils import _validate_input_state
+from qiskit.visualization.utils import _validate_input_state
 if ('ipykernel' in sys.modules) and ('spyder' not in sys.modules):
     try:
         from IPython.core.display import display, HTML
