+++ qiskit/visualization/interactive/_iplot_histogram.py
