+++ qiskit/visualization/interactive/_iplot_qsphere.py
@@ -17,8 +17,8 @@ from string import Template
 
 import numpy as np
 from scipy import linalg
-from qiskit.tools.visualization._utils import _validate_input_state
-from qiskit.tools.visualization.exceptions import VisualizationError
+from qiskit.visualization.utils import _validate_input_state
+from qiskit.visualization.exceptions import VisualizationError
 
 
 if ('ipykernel' in sys.modules) and ('spyder' not in sys.modules):
