+++ qiskit/visualization/interactive/_iplot_state.py
@@ -9,8 +9,8 @@
 Qiskit visualization library.
 """
 import warnings
-from qiskit.tools.visualization.exceptions import VisualizationError
-from qiskit.tools.visualization._utils import _validate_input_state
+from qiskit.visualization.exceptions import VisualizationError
+from qiskit.visualization.utils import _validate_input_state
 from ._iplot_blochsphere import iplot_bloch_multivector
 from ._iplot_cities import iplot_state_city
 from ._iplot_hinton import iplot_state_hinton
