+++ qiskit/visualization/text.py
