+++ test/python/visualization/references/latex_ref.png
