+++ test/python/visualization/references/matplotlib_ref.png
