+++ test/python/visualization/test_circuit_text_drawer.py
@@ -13,8 +13,8 @@ import unittest
 from math import pi
 from codecs import encode
 from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
-from qiskit.tools.visualization import _text as elements
-from qiskit.tools.visualization._circuit_visualization import \
+from qiskit.visualization import text as elements
+from qiskit.visualization.circuit_visualization import \
     _text_circuit_drawer
 from qiskit.test import QiskitTestCase
