+++ test/python/visualization/test_visualization_output.py
